requirements.txt
flask==2.3.2
flask-cors==4.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.1
gtts==2.5.0
gunicorn==20.1.0
annotated-types==0.7.0
flask==2.3.2
flask-cors==4.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.1
gtts==2.5.0
gunicorn==20.1.0
web: gunicorn app:appflask==2.3.2
flask-cors==4.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.1
gtts==2.5.0
gunicorn==20.1.0
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pvporcupine==3.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyttsx3==2.98
requests==2.32.3
rsa==4.9
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
